backend/requirements.txt
fastapi==0.108.0
uvicorn==0.24.0
pydantic==2.5.3
python-dotenv==1.0.0
httpx==0.25.2
deepgram-sdk==2.12.0
pytest==7.4.3
pytest-asyncio==0.23.2
python-multipart==0.0.6
websockets==11.0.3
aiofiles==23.2.1
openai==1.6.1  # For translation capabilities
ffmpeg-python==0.2.0  # For audio processing 